finish migration on ts
