+++ AlexanderMandrov/TypeScript/messenger/src/components/App/App.tsx
@@ -1,42 +1,48 @@
 import React, { useState, useEffect } from 'react';
 import { Route, Switch, useLocation } from 'react-router-dom';
 import { useSelector, useDispatch } from 'react-redux';
+import { Dispatch } from 'redux';
 import './App.scss';
 import { MessageForm } from './MessageForm';
 import { Chat } from './Chat';
 import { Header } from '../Header';
-import Sidebar from '../Sidebar';
+import { Sidebar } from '../Sidebar';
 import { Spinner } from '../Spinner';
-import WelcomePage from '../../pages/WelcomePage';
-import ProfilePage from '../../pages/ProfilePage';
+import { WelcomePage } from '../../pages/WelcomePage';
+import { ProfilePage } from '../../pages/ProfilePage';
 import { Container, Box } from '@material-ui/core';
+import { findChatIdByReceiver, findChatByReceiver } from '../../utils/utils';
+import { AppState } from '../../redux/rootReducer';
 import {
   sendMessage,
   deleteMessage,
   setReceiver,
 } from '../../redux/ducks/chats';
-import { findChatIdByReceiver, findChatByReceiver } from '../../utils/utils';
 
-const App = () => {
-  const [user] = useState('yellso');
+export const App: React.FC = () => {
+  const [user] = useState<string>('yellso');
   let location = useLocation();
 
-  const dispatch = useDispatch();
-  const { chatsReducer } = useSelector((state) => state);
+  const dispatch: Dispatch = useDispatch();
+  const { chatsReducer } = useSelector((state: AppState) => state);
   const { chats, receiver } = chatsReducer;
 
   useEffect(() => {
     if (!receiver) dispatch(setReceiver(location.pathname.split('/')[2]));
   }, []);
 
-  const deleteMessageById = (id) => dispatch(deleteMessage(id));
+  const deleteMessageById: (id: string) => void = (id) =>
+    dispatch(deleteMessage(id));
 
-  const pushMessage = (messageText) => {
-    const editId = findChatIdByReceiver(chats, receiver);
-    dispatch(sendMessage(messageText, user, editId, receiver));
+  const pushMessage: (messageText: string) => void = (messageText) => {
+    if (chats && receiver) {
+      const editId: string = findChatIdByReceiver(chats, receiver);
+      dispatch(sendMessage(messageText, user, editId));
+    }
   };
 
-  const handleChatClick = (receiver) => dispatch(setReceiver(receiver));
+  const handleChatClick: (receiver: string) => void = (receiver) =>
+    dispatch(setReceiver(receiver));
 
   return (
     <>
@@ -57,18 +63,20 @@ const App = () => {
                 path="/chats/:id"
                 render={() => (
                   <>
-                    {!findChatByReceiver(chats, receiver) ? (
+                    {!chats ? (
                       <Box mt={5}>
                         <Spinner />
                       </Box>
                     ) : (
-                      <Chat
-                        user={user}
-                        deleteMessage={deleteMessageById}
-                        getMessageList={() =>
-                          findChatByReceiver(chats, receiver).messages
-                        }
-                      />
+                      receiver && (
+                        <Chat
+                          user={user}
+                          deleteMessage={deleteMessageById}
+                          getMessageList={() =>
+                            findChatByReceiver(chats, receiver).messages
+                          }
+                        />
+                      )
                     )}
                     <MessageForm pushMessage={pushMessage} />
                   </>
@@ -81,5 +89,3 @@ const App = () => {
     </>
   );
 };
-
-export default App;
