+++ AlexanderMandrov/TypeScript/messenger/src/components/Sidebar/index.ts
@@ -1,0 +1,3 @@
+import { Sidebar } from './Sidebar';
+
+export { Sidebar };
