--- AlexanderMandrov/TypeScript/messenger/src/pages/ProfilePage/index.js
@@ -1,3 +1,0 @@
-import ProfilePage from './ProfilePage';
-
-export default ProfilePage;
