+++ AlexanderMandrov/TypeScript/messenger/src/pages/ProfilePage/index.ts
@@ -1,0 +1,3 @@
+import { ProfilePage } from './ProfilePage';
+
+export { ProfilePage };
