+++ AlexanderMandrov/TypeScript/messenger/src/pages/WelcomePage/WelcomePage.tsx
@@ -2,12 +2,10 @@ import React from 'react';
 import './WelcomePage.scss';
 import { Typography } from '@material-ui/core';
 
-const WelcomePage: React.FC = () => {
+export const WelcomePage: React.FC = () => {
   return (
     <Typography className="WelcomePage-title" variant="h5" align="center">
       Select a chat to start messaging
     </Typography>
   );
 };
-
-export default WelcomePage;
