--- AlexanderMandrov/TypeScript/messenger/src/pages/WelcomePage/index.js
@@ -1,3 +1,0 @@
-import WelcomePage from './WelcomePage';
-
-export default WelcomePage;
