+++ AlexanderMandrov/TypeScript/messenger/src/pages/WelcomePage/index.ts
@@ -1,0 +1,3 @@
+import { WelcomePage } from './WelcomePage';
+
+export { WelcomePage };
